--- icedata/datasets/exdark/EXDARK/__init__.py
@@ -1,3 +1,0 @@
-from icedata.datasets.exdark.data import *
-from icedata.datasets.exdark.parser import *
-from icedata.datasets.exdark.dataset import *
